* master:
  svn update
  update addon scripts
  style cleanup
  code cleanup: double promotion warnings
  Node UI: * Made the new custom theme options a bit better and more compact. * Added a few separators to better distinguish these settings from the actual node buttons.

+++ source/blender/editors/transform/transform_conversions.c
@@ -1920,10 +1920,10 @@ static void VertsToTransData(TransInfo *t, TransData *td, TransDataExtension *tx
 		td->val = bweight;
 		td->ival = bweight ? *(bweight) : 1.0f;
 	}
-	else if(t->mode == TFM_SKIN_RESIZE) {
+	else if (t->mode == TFM_SKIN_RESIZE) {
 		MVertSkin *vs = CustomData_bmesh_get(&em->bm->vdata,
-											 eve->head.data,
-											 CD_MVERT_SKIN);
+		                                     eve->head.data,
+		                                     CD_MVERT_SKIN);
 		/* skin node size */
 		td->ext = tx;
 		copy_v3_v3(tx->isize, vs->radius);
@@ -3722,9 +3722,9 @@ void flushTransGraphData(TransInfo *t)
 			switch (sipo->autosnap) {
 				case SACTSNAP_FRAME: /* snap to nearest frame (or second if drawing seconds) */
 					if (sipo->flag & SIPO_DRAWTIME)
-						td2d->loc[0] = (float)(floorf((td2d->loc[0]/secf) + 0.5f) * secf);
+						td2d->loc[0] = floorf((td2d->loc[0] / secf) + 0.5) * secf;
 					else
-						td2d->loc[0] = (float)(floorf(td2d->loc[0]+0.5f));
+						td2d->loc[0] = floorf(td2d->loc[0] + 0.5);
 					break;
 				
 				case SACTSNAP_MARKER: /* snap to nearest marker */
